requirements.txt
fastapi
pydantic
uvicorn[standard]
prometheus-client
kafka-python
